requirements.txt
auto_mix_prep==0.2.0
hydra-core==1.3.2
matplotlib==3.3.3
numpy==1.19.4
omegaconf==2.3.0
opencv_python==4.7.0.72
pandas==1.1.5
Pillow==10.0.0
pytorch_model_summary==0.1.2
pytransform3d==2.3.1
scikit_image==0.17.2
scikit_learn==1.3.0
scipy==1.5.4
skimage==0.0
timm==0.6.12
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
transformations==2021.6.6
